requirements.txt
django==1.8
psycopg2==2.5.1
Pillow
django-tinymce
sorl-thumbnail
django-treebeard
django-filebrowser-no-grappelli
django-simple-captcha
django-bootstrap3
django-mailer-2
django-request-provider
django-contrib-comments
django-newsletter
reportlab==2.7
html5lib==1.0b8
xhtml2pdf==0.0.6
qrcode
pytils-0.3

#reportlab==2.7
#-e git+git://github.com/darklow/django-suit.git@v.0.2.5#egg=django-suit
#django-registration-redux
#simplejson==3.3.0
#django-cms
#django-filer
#django-suit-ckeditor
#django-debug-toolbar
#Flup==1.0.2
